% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_brier_test.R
\name{bss}
\alias{bss}
\title{Calculate Brier Skill Score}
\usage{
bss(bs_forecast, bs_reference)
}
\arguments{
\item{obj}{: coefficient predicted for a survival model e.g. with glmnet \cr
test_data: a test holdout dataframe, is recommended to test the model in a separate dataset from which was used for training\cr}
}
\value{
predicted linear predictor for each individual in the test set
}
\description{
Estimate Brier Skill Score usually from reference
}
