% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_brier_test.R
\name{fun_ibrier_score_reference}
\alias{fun_ibrier_score_reference}
\title{Integrated Brier Score}
\usage{
fun_ibrier_score_reference(brier)
}
\arguments{
\item{obj}{: survival model object for example a glmnet fit. \cr
coef: coefficient beta is the default. \cr
new_data: a test holdout dataframe default test\cr
train_data : train dataframe default train\cr}
}
\value{
a coxph fit object
}
\description{
Predict integration Brier Score in a new dataset.
}
