% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_brier_test.R
\name{pred_lp}
\alias{pred_lp}
\title{Calculate Linear Predictor}
\usage{
pred_lp(obj, test_data, time = os_months, status = os_deceased, iter = 0,
  ...)
}
\arguments{
\item{obj}{: coefficient predicted for a survival model e.g. with glmnet \cr
test_data: a test holdout dataframe, is recommended to test the model in a separate dataset from which was used for training\cr}
}
\value{
predicted linear predictor for each individual in the test set
}
\description{
Predict linear components or hazard ratio.
}
