% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_brier_test.R
\name{pred_surv_prob}
\alias{pred_surv_prob}
\title{Predict Surv Prob}
\usage{
pred_surv_prob(survdata, time = time, status = status, coeff = covariates,
  iter = 0, inits = 0, ...)
}
\arguments{
\item{obj}{: coefficient predicted for a survival model e.g. with glmnet \cr
test_data: a test holdout dataframe, is recommended to test the model in a separate dataset from which was used for training\cr}
}
\value{
predicted linear predictor for each individual in the test set
}
\description{
Estimate survival ROC, option to have the integrated survival ROC.
}
