% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_surv_manipulation.R
\name{stick_covar}
\alias{stick_covar}
\title{Stick covariates to a dataframe for subject}
\usage{
stick_covar(survdata, covars)
}
\arguments{
\item{survdata:}{survival dataset \cr
covars : covariates \cr}
}
\value{
predicted linear predictor for each individual in the test set
}
\description{
Sitck covars...
}
